+++ ydb/core/persqueue/common/sourceid_info.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
-#include <ydb/core/persqueue/heartbeat.h>
+#include "heartbeat.h"
 #include <ydb/core/protos/pqconfig.pb.h>
 
 namespace NKikimr::NPQ {
