+++ ydb/core/persqueue/events/internal.h
@@ -9,7 +9,7 @@
 #include <ydb/core/persqueue/common/metering.h>
 #include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
 #include <ydb/core/persqueue/public/counters/percentile_counter.h>
-#include <ydb/core/persqueue/sourceid_info.h>
+#include <ydb/core/persqueue/common/sourceid_info.h>
 #include <ydb/core/persqueue/public/write_id.h>
 #include <ydb/core/tablet/tablet_counters.h>
 #include <ydb/library/persqueue/topic_parser/topic_parser.h>
